+++ packages/plugin-pha/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.1.1
 
+- [feat] support resource_prefetch to preload resource
+- [feat] support enableExpiredManifest
 - [feat] support configure `enableExpiredManifest`
 
 ## 1.1.0